requirements.txt
Flask==2.3.2
google-cloud-storage==2.9.0
pickleshare==0.7.5
numpy==1.23.5
scikit-learn==1.2.2
pandas==1.2.5